added S3 as minio for test purpose
