+++ save-backend/src/test/kotlin/com/saveourtool/save/backend/controller/JpaSpecificationTest.kt
@@ -3,7 +3,7 @@ package com.saveourtool.save.backend.controller
 import com.saveourtool.save.agent.AgentState
 import com.saveourtool.save.backend.configs.ApplicationConfiguration
 import com.saveourtool.save.backend.repository.AgentStatusRepository
-import com.saveourtool.save.backend.utils.MySqlExtension
+import com.saveourtool.save.backend.utils.InfraExtension
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -15,7 +15,7 @@ import org.springframework.context.annotation.Import
 @Import(ApplicationConfiguration::class)
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@ExtendWith(MySqlExtension::class)
+@ExtendWith(InfraExtension::class)
 class JpaSpecificationTest {
     @Autowired
     lateinit var agentStatusRepository: AgentStatusRepository
