+++ save-backend/src/test/kotlin/com/saveourtool/save/backend/controller/UsersControllerTest.kt
@@ -1,7 +1,7 @@
 package com.saveourtool.save.backend.controller
 
 import com.saveourtool.save.backend.SaveApplication
-import com.saveourtool.save.backend.utils.MySqlExtension
+import com.saveourtool.save.backend.utils.InfraExtension
 import com.saveourtool.save.domain.Role
 import com.saveourtool.save.entities.OriginalLogin
 import com.saveourtool.save.entities.User
@@ -20,7 +20,7 @@ import org.springframework.test.web.reactive.server.expectBody
 
 @SpringBootTest(classes = [SaveApplication::class])
 @AutoConfigureWebTestClient
-@ExtendWith(MySqlExtension::class)
+@ExtendWith(InfraExtension::class)
 class UsersControllerTest {
     @Autowired
     private lateinit var webTestClient: WebTestClient
