+++ save-backend/src/test/resources/application.properties
@@ -7,11 +7,11 @@ backend.demoUrl=stub
 backend.initialBatchSize=20
 backend.fileStorage.location=files
 # s3
-backend.s3-storage.endpoint=stub
-backend.s3-storage.bucketName=stub
-backend.s3-storage.prefix=stub
-backend.s3-storage.credentials.accessKeyId=stub
-backend.s3-storage.credentials.secretAccessKey=stub
+backend.s3-storage.endpoint=http://localhost:9000
+backend.s3-storage.bucketName=cnb
+backend.s3-storage.prefix=
+backend.s3-storage.credentials.accessKeyId=admin
+backend.s3-storage.credentials.secretAccessKey=12345678
 management.endpoints.web.exposure.include=health,info,prometheus
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
 spring.liquibase.change-log=db/db.changelog-master.xml
