+++ save-backend/src/test/resources/docker-compose.yaml
@@ -1,0 +1,23 @@
+version: '3.9'
+services:
+  minio:
+    image: minio/minio:latest
+    command: server /data --console-address ":9090"
+    ports:
+      - 9000:9000
+      - 9090:9090
+    environment:
+      MINIO_ROOT_USER: admin
+      MINIO_ROOT_PASSWORD: 12345678
+
+  minio-create-bucket:
+    image: minio/mc:latest
+    depends_on:
+      - minio
+    entrypoint:
+      - /bin/sh
+      - -c
+      - |
+        /usr/bin/mc alias set minio http://minio:9000 admin 12345678
+        /usr/bin/mc mb --ignore-existing minio/cnb
+        /usr/bin/mc policy set public minio/cnb
